requirement.txt
asgiref==3.6.0
Django==4.1.5
Django-Verify-Email==2.0.3
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2022.7
